--- manual.txt
@@ -1,2 +1,0 @@
-OK
-Them dong moi ne123
