+++ public/xuly.js
@@ -1,7 +1,8 @@
 
-var socket = io("52.146.10.164");
+var socket = io("localhost:80");
 
 socket.on("server-send-files",function(data){
+    alert(111);
     for(let i=0;i<data.length;i++){
         let link = "/image/"+data[i];
 
@@ -9,14 +10,14 @@ socket.on("server-send-files",function(data){
     }
 });
 socket.on("server-send-new-file",function(data){
-
+    alert(222);
     let name = "./image/" + data;
     $("#listImage").append("<a target ='_blank' href="+name+"><img class='img' src="+name+"></a>");
 });
 
 $(document).ready(function(){
     socket.emit("client-get-files");
-    $("#a1").mousedown(function(ev){
+    $("#a1g").mousedown(function(ev){
         if(ev.which == 3)
         {
               alert("Right mouse button clicked on element with id myId");
